+++ ale/base/data_naif.py
@@ -3,8 +3,7 @@ import spiceypy as spice
 import warnings
 from concurrent.futures import ThreadPoolExecutor
 from concurrent.futures import as_completed
-import logging
-import json 
+import json
 import os
 
 import numpy as np
@@ -44,7 +43,7 @@ class NaifSpice():
         elif isinstance(self.kernels, dict) and not self.use_web and not self.search_kernels:
             self.kset = pyspiceql.KernelSet(self.kernels)
         elif not self.use_web:
-            logger.warn("No kernels were specified. No kernels will be loaded.")
+            logger.warning("No kernels were specified. No kernels will be loaded.")
         return self
 
     def __exit__(self, exc_type, exc_val, exc_tb):
@@ -115,8 +114,9 @@ class NaifSpice():
                 if spice_root:
                     search_results = kernel_access.get_metakernels(spice_root, missions=self.short_mission_name, years=self.utc_start_time.year, versions='latest')
                     if search_results['count'] != 0:
-                        self._kernels = [search_results['data'][0]['path']]
-                        logger.debug(f"Found metakernel: {self._kernels}")
+                        logger.debug(f"Found metakernel: {search_results['data'][0]['path']}")
+                        self._kernels = kernel_access.get_kernels_from_metakernel(search_results['data'][0]['path'])
+                        logger.debug(f"Retrieved kernels: {self._kernels}")
 
                 if self._kernels == {}:
                     logger.debug("Failed to find metakernels, falling back to SpiceQL")