+++ CHANGELOG.md
@@ -35,6 +35,9 @@ release.
 
 ## [Unreleased]
 
+### Fixed
+- Auto-correct stale `/usgs/cpkgs/isis3/data` paths in metakernels using ALESPICEROOT or ISISDATA. [#703](https://github.com/DOI-USGS/ale/pull/703)
+
 ## [1.2.0] - 2026-05-20
 
 ### Added